pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pbrecipe"
version = "0.1.0"
description = "Recipe management application with web export"
readme = "README.md"
license = { file = "LICENSE" }
authors = [{ name = "Philippe Poilbarbe", email = "philippe@cardolan.net" }]
keywords = ["recipes", "cooking", "cocktails", "gui", "qt"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Environment :: X11 Applications :: Qt",
    "Intended Audience :: End Users/Desktop",
    "Topic :: Utilities",
]
requires-python = ">=3.12"
dependencies = [
    "PySide6>=6.6.0",
    "ruamel.yaml>=0.18",
    "SQLAlchemy>=2.0",
    "PyMySQL>=1.1",
    "psycopg2>=2.9",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4",
    "pre-commit>=3.7",
    "pytest>=8",
    "pytest-qt>=4",
    "pytest-cov>=5",
]

[project.scripts]
pbrecipe = "pbrecipe.__main__:main"

[project.urls]
Homepage   = "https://github.com/ppoilbarbe/PBRecipe"
Repository = "https://github.com/ppoilbarbe/PBRecipe"
Issues     = "https://github.com/ppoilbarbe/PBRecipe/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/pbrecipe"]

[tool.hatch.build.targets.wheel.sources]
"src/pbrecipe/resources" = "pbrecipe/resources"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts   = ["--cov=pbrecipe", "--cov-report=term-missing"]

[tool.coverage.run]
source = ["src"]

[tool.ruff]
line-length    = 88
target-version = "py312"
src            = ["src"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP"]
extend-select = ["PGH", "BLE"]
ignore = [
    "N802",   # Qt override methods must match exact C++ names (resizeEvent, closeEvent, …)
]

[tool.ruff.format]
quote-style               = "double"
indent-style              = "space"
skip-magic-trailing-comma = false
line-ending               = "lf"
